Refactoring code to address style issues
This is primarily focused on readability, but as a side effect some of
the changes should additionally increase performance.

Running the tests before and after the change, I experienced an
improvement in test speed of about 4.5% which. This may or may not show
in real world testing, but it does indicate that there is not a decrease
in performance while adjusting style to be more standard.